Some update on action response related types
