+++ app/definitions/types.ts
@@ -96,23 +96,3 @@ export interface PollVoteAnswer {
   readonly pollQuestionId: string;
   readonly optionIds: string[];
 }
-
-// Action response
-export interface ActionSuccess<DataT = undefined> {
-  readonly success: true;
-  readonly data: DataT;
-  readonly error?: undefined;
-}
-
-export interface ActionFailure {
-  readonly success: false;
-  readonly data?: undefined;
-  readonly error: {
-    code: string;
-    message?: string | undefined;
-  };
-}
-
-export type ActionResponse<DataT = undefined> =
-  | ActionSuccess<DataT>
-  | ActionFailure;
